requirements.txt
anyio==3.6.1
asgiref==3.5.2
astroid==2.9.3
click==8.1.3
coverage==6.3.2
fastapi==0.78.0
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.2
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
pylint==2.12.2
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
wrapt==1.13.3
alembic==1.7.7
Werkzeug==2.1.2
pyjwt==2.1.0